loan-company/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.4
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
URLObject==2.4.3